Added engine parameter to game
Because game handles input actions and translates them to functions, it needs access to the engine so that when in a "Quit Application" state, it actually has a way to quit the application (called "engine_stop(game->engine)"). Without this, the only way to exit would be to always press the close button on the window.