Added basic keybind support
Keybinds are now capable of being defined and set in the program. However, the config load and save functions still need to be defined.

I am not too happy with the amount of manual upkeep required here, but it's probably the biggest job of the input and config to be upkept anyway, so I'll accept it until a better solution comes along.