System overhaul in regards to input actions
Moved input_action into input, and let game ask input directly for its state. Game is also now in charge of stopping the engine when a quit action is set.

This clears up responsibilities, as game really ought to be in charge of managing state and behaviour. It also clarifies world's purpose as being intended as a player's save data.