Added new game and pause menu states
Also includes an update to the main menu, and a renaming of the open pause menu keybind and action to toggle pause menu. Pause menu returns to the main menu.

Renderer process has been updated to accommodate changes, whereby menus GUI elements are now capable of clearing parts of the rendered game to use as background (BearLibTerminal only supports backgrounds on Layer 1).