Added bool prompt, new state, and color tool
Bool prompt is a dialog box with a yes/no answer and a question. Currently, it appears when quit to main menu is selected in the pause menu.

The new state is for the pause menu quit option, and highlights how substates should work (take advantage of the state stack.

All rendering now takes place on one layer (layer 0), which makes it easier to add effects when necessary.

On effects, there is now a color tool, which currently allows for blending (adding or subtracting) colors. This is useful for layering effects, such as when a prompt appears, and a darker or lighter background would help make it stand out.