Added config load and save
If no config.ini file is found when config_load is called, a new file with the default config settings is created. If such a file exists, config will currently attempt to set the keybindings.

Additionally, added support for holding alt keys as part of a keybind, and renamed "alt_key" to the more appropriate "key_alternative".