Added support for game saves and loading directories
Config now stores a file directory (eventually will be a folder as stated) for where to save a game. This gets loaded when the app launches, so users can move their saves around with ease.

The world can now also be saved out, and is saved out as a binary. Currently the whole world structure gets dumped out (or fully loaded in). So the current expectation is that everything in the world is "saveable".

This could change in the future, as the world has explicit world_save() and world_load() functions, intended to handle what gets sent out.