Final day
Push for the results of the final day.

Attempted to locate a bug, whereby the application sometimes doesn't open, and often doesn't close. The bug seems to be coming from either iniparser (hangs on `iniparse_deldict()`), or from some kind of interaction involving the memory of iniparser (bad pointer).

Regardless, this was the final result, pushed for future reading and research.